poke_api/pyproject.toml
[tool.poetry]
name = "poke-api"
version = "0.1.0"
description = ""
authors = ["Theodore Bartemes <taybartemes@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.2.3"
requests = "^2.28.2"
pillow = "^9.5.0"
rich = "^13.3.3"
termcolor = "^2.2.0"
opencv-python = "^4.7.0.72"
asciichartpy = "^1.5.25"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pytest = "^7.2.2"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
